Formatting, cleaning, commenting
